requirements.txt
# Web Framework
flask==3.0.0
flask-cors==4.0.0

# Task Queue
celery==5.3.4
redis==5.0.1

# OpenAI API (for AI-powered translation and transcription)
openai==0.28.1  # Temporarily reverted for compatibility

# AssemblyAI for advanced transcription with speaker diarization
assemblyai==0.33.0

# HTTP clients
httpx==0.24.1  # Pin for OpenAI v0.28.1 compatibility
requests>=2.31.0  # Used for RapidAPI calls

# ElevenLabs API (for Georgian TTS)
elevenlabs==2.23.0

# Google Cloud TTS with Gemini (alternative TTS provider)
google-cloud-texttospeech>=2.29.0

# Video Download (fallback when RapidAPI not available)
yt-dlp>=2024.11.18

# Audio/Video Processing
ffmpeg-python==0.2.0
pydub==0.25.1

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Cloud Storage (Cloudflare R2)
boto3==1.34.0

# Production Server
gunicorn==21.2.0

# Logging
structlog==23.2.0
python-json-logger==2.0.7

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1
